# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_booking.BoardResponse:
    properties:
      bookings:
        items:
          $ref: '#/definitions/internal_booking.BookingResponse'
        type: array
      business_date:
        type: string
      closed:
        description: วันปิดประจำสัปดาห์
        type: boolean
      closes_at:
        type: string
      opens_at:
        type: string
      seats:
        type: integer
      slots:
        items:
          $ref: '#/definitions/internal_booking.BoardSlot'
        type: array
    type: object
  internal_booking.BoardSlot:
    properties:
      booked:
        description: คนในร้านสูงสุดในช่วงนี้
        type: integer
      end_at:
        type: string
      start_at:
        type: string
    type: object
  internal_booking.BookingRequest:
    properties:
      date:
        example: "2026-10-10"
        type: string
      end_time:
        example: "19:00"
        type: string
      party_size:
        example: 2
        type: integer
      restaurant_id:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
      start_time:
        example: "18:00"
        type: string
    required:
    - date
    - end_time
    - party_size
    - start_time
    type: object
  internal_booking.BookingResponse:
    properties:
      business_date:
        description: วันทำการของรอบที่ booking นี้อยู่ (ใช้ทำป้าย "(เช้าวันที่ n)")
        type: string
      can_change:
        description: แก้/ยกเลิกได้ตอนนี้ไหม (หน้าเว็บใช้ disable ปุ่มพร้อมเหตุผล)
        type: boolean
      cancel_until:
        type: string
      cancelled_at:
        type: string
      code:
        example: JY-7F3K2A
        type: string
      created_at:
        type: string
      customer_name:
        type: string
      end_at:
        type: string
      id:
        type: string
      party_size:
        type: integer
      restaurant:
        $ref: '#/definitions/internal_booking.RestaurantSummary'
      start_at:
        type: string
      status:
        example: active
        type: string
    type: object
  internal_booking.RestaurantSummary:
    properties:
      address:
        type: string
      id:
        type: string
      name:
        type: string
      overnight:
        type: boolean
    type: object
...
  jongyoung_internal_httputil.ErrorBody:
    properties:
      code:
        example: NOT_ENOUGH_SEATS
        type: string
      details: {}
      message:
        example: ที่นั่งไม่พอ
        type: string
    type: object
  jongyoung_internal_httputil.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/jongyoung_internal_httputil.ErrorBody'
    type: object
host: api.jongyoung.localhost
info:
  contact: {}
  description: ระบบจองโต๊ะร้านอาหาร + รีวิว — เวลาทั้งหมดเป็น timestamp เต็ม (RFC
    3339) ส่วน date = วันทำการ
  title: jongyoung API
  version: "1.0"
paths:
  /bookings:
    post:
      consumes:
      - application/json
      description: 409 NOT_ENOUGH_SEATS = ที่นั่งไม่พอ (details.available, details.at);
        409 DUPLICATE_BOOKING = มีการจองที่ทับช่วงนี้อยู่แล้ว (details.booking_id)
      parameters:
      - description: วันทำการ + เวลา + จำนวนคน
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_booking.BookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_booking.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: จองโต๊ะ
      tags:
      - bookings
  /bookings/{id}:
    delete:
      parameters:
      - description: booking id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: CANCEL_WINDOW_PASSED (details.cancel_until) หรือไม่ใช่การจองของตัวเอง
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ยกเลิกการจอง (เปลี่ยนสถานะเป็น cancelled ไม่ลบจริง)
      tags:
      - bookings
    get:
      parameters:
      - description: booking id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_booking.BookingResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: รายละเอียดการจอง (หน้ายืนยัน) — เจ้าของการจองหรือเจ้าของร้านเท่านั้น
      tags:
      - bookings
    put:
      consumes:
      - application/json
      parameters:
      - description: booking id
        in: path
        name: id
        required: true
        type: string
      - description: ไม่ต้องส่ง restaurant_id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_booking.BookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_booking.BookingResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: แก้ไขการจอง (จำนวนคน/วัน/เวลา) — ต้องยังไม่เลยเส้นตายของเวลาเดิม
      tags:
      - bookings
...
  /me/bookings:
    get:
      parameters:
      - description: upcoming (ค่าเริ่มต้น) | past | cancelled
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_booking.BookingResponse'
            type: array
      security:
      - BearerAuth: []
      summary: การจองของฉัน
      tags:
      - me
...
  /restaurants/{id}/bookings:
    get:
      parameters:
      - description: restaurant id
        in: path
        name: id
        required: true
        type: string
      - description: วันทำการ YYYY-MM-DD — แขกตี 1 คืนวันเสาร์อยู่ในวันเสาร์
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_booking.BoardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jongyoung_internal_httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: บอร์ดการจองรายวันทำการของร้าน (เจ้าของร้านเท่านั้น) + คนในร้านต่อช่วง
        30 นาที
      tags:
      - restaurants
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: พิมพ์ "Bearer " ตามด้วย access token จาก Keycloak
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"